Made player classes work removed some code from gb
