+++ player.py
@@ -7,24 +7,33 @@ Professor: A. Nuzen
 </pre>
 """
 
+from cell import cell
 class player(object):
     def __init__(self, playerNum, score:int, is_turn:bool)->None:
         self.playerNum = playerNum
         self.score = score
         self.is_turn = is_turn
-        self.is_human = True
+        
+        shipCells = []
+        for r in range(10):
+            row = []
+            for c in range(10):
+                row.append(cell(False, False, "NA"))
+            shipCells.append(row)
 
-    def make_hit(self)->None:
-        #Need to implement
-        return
+        attackingCells = []
+        for r in range(10):
+            row = []
+            for c in range(10):
+                row.append(cell(False, False, "NA"))
+            attackingCells.append(row)
+
+        self.shipCells = shipCells
+        self.attackingCells = attackingCells
+        self.numOfHits = 0
+
+    def incrementHits(self):
+        self.numOfHits += 1
 
     def __str__(self):
         return ("Player %d", self.playerNum)
-
-    def setHits(self, ship_types):
-        self.hit = {}
-        self.ship_size = {}
-        for item in ship_types:
-            print(item.name)
-            self.ship_size[item.name] = item.length
-            self.hit[item.name] = 0
